python-ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
openai==1.57.0
python-dotenv==1.0.1
pydantic==2.9.2
numpy==2.1.3
scikit-learn==1.5.2
httpx==0.27.0
